+++ plugins/clickhouse/src/components/panel/VisualizationChart.tsx
@@ -6,19 +6,17 @@ import VisualizationChartPie from './VisualizationChartPie';
 
 interface IVisualizationChartProps {
   chart: string;
-  operation: string;
   data: IVisualizationData[];
 }
 
 const VisualizationChart: React.FunctionComponent<IVisualizationChartProps> = ({
   chart,
-  operation,
   data,
 }: IVisualizationChartProps) => {
   if (chart === 'bar') {
-    return <VisualizationChartBar operation={operation} data={data} />;
+    return <VisualizationChartBar data={data} />;
   } else if (chart === 'pie') {
-    return <VisualizationChartPie operation={operation} data={data} />;
+    return <VisualizationChartPie data={data} />;
   } else {
     return null;
   }